+++ vue/src/store/index.js
@@ -4,6 +4,7 @@ import MainModel from "@/store/modules/MainModel";
 import login from "@/store/modules/login";
 import iframeModel from "@/store/modules/iframeModel";
 import registration from "@/store/modules/registration";
+import filters from "@/store/modules/filters";
 
 export default createStore({
     state: {},
@@ -14,6 +15,7 @@ export default createStore({
         registration,
         iframeModel,
         MainModel,
+        filters,
         models,
         login,
     },
